chore: prepping multiple binaries
